+++ Core/Vector.cl
@@ -1,14 +1,14 @@
 #ifndef VECTOR_CL
 #define VECTOR_CL
 
-double4 vector_getUnitary(double4 vector)
+float4 vector_getUnitary(float4 vector)
 {
-    double vectorLength = length(vector);
+    float vectorLength = length(vector);
 
-    if (isequal(vectorLength, 0.0))
-        return (double4)0;
+    if (isequal(vectorLength, (float)0.0))
+        return (float4)0;
 
-    double4 unitary = vector / vectorLength;
+    float4 unitary = vector / vectorLength;
 
     return unitary;
 }
